feat: Set up the graph-api project.
Its main functionalities include: connection with a graph database and
implementation of the basic queries for interacting with the db. It also
implements a logging mechanism and supports dockerization.

The required documents are passed as input through configuration files and
the API is implemented in FastAPI.

Besides README.md, you can find: CHANGELOG.md for reporting the changes of
each pushed version, CONTRIBUTING.md for listing the rules for contributing
to the project and TODO.md to have a glance at the tasks that have to be done.